refactor: ZeroG finish
