+++ .gitignore
@@ -20,11 +20,3 @@ __pycache__
 /.vscode
 .vscode
 /common/apikey.py
-/LLMPredictor/test.py
-LLMZeroshot/test.py
-environment.py
-LLMZeroshot/ZeroGv2/logs/\['cora']/arxiv_citeseer_pubmed.txt
-LLMZeroshot/ZeroGv2/arxiv_index.pkl
-LLMZeroshot/ZeroGv2/logs/\['photo']/computer.txt
-LLMZeroshot/ZeroGv2/logs/\['photo']/computer.txt
-LLMZeroshot/ZeroGv2/sample_analysis.py
