+++ common/descriptions.py
@@ -62,103 +62,14 @@ IT_DESC = {
 
 
 ZeroG_DESC = {
-    "computer": "This dataset represents a subgraph of an e-commerce network focused on computer-related items. Each node corresponds to an item in the Computer category, with node features consisting of reviews for each item. Edges denote co-purchase or co-review relationships between items. The prediction task involves classifying each item into one of 10 sub-categories, including Computer Accessories & Peripherals, Tablet Accessories, Laptop Accessories, Computers & Tablets, Computer Components, Data Storage, Networking Products, Monitors, Servers, and Tablet Replacement Parts.",
-    "photo": "This dataset represents a subgraph of an e-commerce network focused on photo-related items. Each node denotes an item in the Photo category, with node features consisting of reviews for each item. Edges denote co-purchase or co-review relationships between items. The prediction task involves classifying each item into one of 12 sub-categories, including: Video Surveillance, Accessories, Binoculars & Scopes, Video, Lighting & Studio, Bags & Cases, Tripods & Monopods, Flashes, Digital Cameras, Film Photography, Lenses, and Underwater Photography",
     "cora": "The Cora dataset is a fundamental resource in the field of graph learning, particularly within the realm of machine learning research. It represents a network of scientific publications. There are 7 categories in Cora: Theory: This category covers theoretical aspects of machine learning and AI. Reinforcement Learning: This category includes research on reinforcement learning, a type of machine learning where an agent learns to make decisions to achieve a goal, focusing on algorithms, methodologies, and applications in decision-making areas. Genetic Algorithms: This category deals with genetic algorithms, a type of optimization algorithm inspired by natural evolution. Neural Networks: This category focuses on artificial neural networks, a subset of machine learning mimicking the human brain, covering various architectures, training techniques, and applications. Probabilistic Methods: This category pertains to research on probabilistic methods in machine learning, using probability mathematics to handle uncertainty and make predictions. Case Based: This category focuses on case-based reasoning in AI, a method that solves new problems by referring to similar past cases. Rule Learning: This category is about rule-based learning in machine learning, involving the generation of rules for decision-making systems, focusing on algorithms, transparency, and applications in fields requiring interpretability. The average degree of Cora is 4.",
     "citeseer": "The Citeseer dataset is a prominent academic resource in the field of computer science, categorizing publications into six distinct areas. These are Agents, focusing on intelligent agents; Machine Learning (ML), covering all aspects of learning techniques and applications; Information Retrieval (IR), dealing with data and text indexing and retrieval; Databases (DB), related to database management and data mining; Human-Computer Interaction (HCI), emphasizing computer technology interfaces for humans; and Artificial Intelligence (AI), a broad category encompassing general AI theory and applications, excluding certain subfields. The average degree of this graph is 2.",
     "pubmed": "The PubMed dataset comprises three categories: Experimental studies on diabetes mechanisms and therapies, Type 1 Diabetes research focusing on autoimmune processes and treatments, and Type 2 Diabetes studies emphasizing insulin resistance and management strategies. Each category addresses specific aspects of diabetes research, aiding in understanding and treating this complex disease. The average degree of this graph is 4.5.",
     "arxiv": "The arXiv dataset is a notable resource in the field of graph learning, particularly in the area of computer science research. This dataset forms a directed graph representing the citation network among all Computer Science papers on arXiv, as indexed by the Microsoft Academic Graph (MAG). Each node in this network corresponds to a paper, and directed edges indicate citations. The dataset's primary challenge is predicting the 40 subject areas of arXiv CS papers, such as cs.AI, cs.LG, and cs.OS. The task is structured as a 40-class classification problem.",
     "wikics": "The Wiki CS dataset is a comprehensive collection of Wikipedia entries, systematically categorized into ten distinct areas of computer science. These categories include Computational Linguistics, focusing on the intersection of computer science and linguistics; Databases, covering database technologies and theories; Operating Systems, detailing the software that manages computer hardware; Computer Architecture, exploring the design and structure of computer systems; Computer Security, addressing the protection of information systems; Internet Protocols, discussing the rules governing internet data exchange; Computer File Systems, about methods for storing and organizing computer files; Distributed Computing Architecture, concerning computations spread across multiple machines; Web Technology, focusing on the technologies underpinning the web; and Programming Language Topics, which includes various aspects of programming languages. This dataset serves as a valuable resource for understanding diverse computer science topics as represented in Wikipedia, reflecting the breadth and depth of the field.",
-    "reddit":"Reddit is also a social network where each node denotes a user, the node features are the content of users' historically published subreddits, and edges denote whether two users have replied to each other. The prediction task is to classify whether a user is in the top 50% popular (average score of all subreddits).",
-    "instagram":"Instagram is a social network where edges represent following relationships, nodes represent users, and the prediction task is to classify commercial and normal users in this network.",
-}
-
-
-ZeroG_LABEL_DESC = {
-    "cora": {
-        'Rule_Learning': "The 'Rule Learning' category pertains to research papers that concentrate on the domain of rule-based learning, also known as rule-based machine learning. Rule learning is a method in machine learning that involves the generation of a set of rules to predict the output in a decision-making system based on the patterns discovered from the data. These rules are often in an 'if-then' format, making them interpretable and transparent. This category would encompass research involving various rule learning algorithms, their enhancements, theoretical foundations, and applications. Rule learning methods are particularly beneficial in domains where interpretability and understanding of the learned knowledge is important, such as in medical diagnosis, credit risk prediction, and more.",
-        'Neural_Networks': "The 'Neural Networks' category refers to research papers revolving around the concept of artificial neural networks (ANNs). Neural networks are a subset of machine learning algorithms modelled after the human brain, designed to 'learn' from observational data. They are the foundation of deep learning technologies and can process complex data inputs, find patterns, and make decisions. The network consists of interconnected layers of nodes, or 'neurons', and each connection is assigned a weight that shapes the data and helps produce a meaningful output. Topics covered under this category could range from the architecture and function of different neural network models, advancements in training techniques, to their application in a multitude of fields such as image and speech recognition, natural language processing, and medical diagnosis.", 
-        'Case_Based': "The 'Case Based' category refers to research papers focusing on case-based reasoning (CBR) in the field of artificial intelligence. Case-based reasoning is a problem-solving approach that utilizes specific knowledge of previously encountered, concrete problem situations (cases). In this method, a new problem is solved by finding similar past cases and reusing them in the new situation. The approach relies on the idea of learning from past experiences to solve new problems, which makes it relevant in many applications including medical diagnosis, legal decision-making, and others. Thus, the 'Case Based' category would include papers that primarily focus on this particular methodology and its various aspects.", 
-        'Genetic_Algorithms': "The 'Genetic Algorithms' category would include research papers related to genetic algorithms (GAs). Genetic algorithms are a type of optimization and search algorithms inspired by the process of natural selection and genetics. These algorithms generate solutions to optimization problems using techniques inspired by natural evolution, such as inheritance, mutation, selection, and crossover. In practice, genetic algorithms can be used to find solutions to complex problems that are difficult to solve with traditional methods, particularly in domains where the search space is large, complex, or poorly understood. This category would cover various aspects of genetic algorithms, including their design, analysis, implementation, theoretical background, and diverse applications.", 
-        'Theory': "The 'Theory' category likely refers to research papers that delve into the theoretical aspects of machine learning and artificial intelligence. This includes a broad array of topics such as theoretical foundations of various machine learning algorithms, performance analysis, studies on learning theory, statistical learning, information theory, and optimization methods. Additionally, it could encompass the development of new theoretical frameworks, investigations into the essence of intelligence, the potential for artificial general intelligence, as well as the ethical implications surrounding AI. Essentially, the 'Theory' category encapsulates papers that primarily focus on theoretical concepts and discussions, contrasting with more application-oriented research which centers on specific techniques and their practical implementation.", 
-        'Reinforcement_Learning': "The 'Reinforcement Learning' category refers to research papers focusing on the area of machine learning known as reinforcement learning (RL). Reinforcement learning is a type of machine learning where an agent learns to make decisions by taking actions in an environment to achieve a goal. The agent learns from the consequences of its actions, rather than from being explicitly taught, and adjusts its behavior based on the positive or negative feedback it receives, known as rewards or penalties. This category would include research exploring various RL algorithms, methodologies, theoretical underpinnings, performance enhancements, and practical applications. This field is particularly relevant in areas where decision making is crucial, such as game playing, robotics, resource management, and autonomous driving.", 
-        'Probabilistic_Methods': "The 'Probabilistic Methods' category pertains to research papers that focus on probabilistic methods and models in machine learning and artificial intelligence. Probabilistic methods use the mathematics of probability to make predictions and decisions. They provide a framework to handle and quantify the uncertainty and incomplete information, which is a common scenario in real-world problems. This category could include topics like Bayesian networks, Gaussian processes, Markov decision processes, and statistical techniques for prediction and inference. These methods have applications in various areas such as computer vision, natural language processing, robotics, and data analysis, among others, due to their ability to model complex, uncertain systems and make probabilistic predictions."
-    }, 
-    "citeseer": {
-        'Agents': 'Software entities that perform tasks autonomously, often interacting with other agents and systems.', 
-        'ML (Machine Learning)': 'A field of AI focused on algorithms that improve automatically through experience and data.', 
-        'IR (Information Retrieval)': 'The process of obtaining relevant information from large collections of data, often with search engines.', 
-        'DB (Databases)': 'Organized collections of structured information stored electronically, accessible in various ways.', 
-        'HCI (Human-Computer Interaction)': 'The study of how people interact with computers and designing interfaces that facilitate this.', 
-        'AI (Artificial Intelligence)': 'The simulation of human intelligence processes by machines, especially computer systems.'
-    },
-    "pubmed": {
-        'Experimental': 'Pertaining to a trial or test phase involving innovative methods or studies.', 
-        'Diabetes Mellitus Type 1': "The PubMed literature category 'Diabetes Mellitus Type 1' focuses on scientific research related specifically to Type 1 diabetes mellitus. ", 
-        'Diabetes Mellitus Type 2': "The PubMed category 'Diabetes Mellitus Type 2' focuses on research related to Type 2 diabetes (T2D). "
-    }, 
-    "wikics": {
-        'Computational Linguistics': 'The study of using computational methods to process and analyze human language data.', 
-        'Databases': 'Structured collections of data, typically managed by database management systems to enable retrieval and manipulation.', 
-        'Operating Systems': 'Software that manages computer hardware and software resources, providing common services for application programs.', 
-        'Computer Architecture': "The design and organization of a computer's core components, including the CPU, memory, and input/output systems.", 
-        'Computer Security': 'Measures and protocols for protecting computer systems and data from unauthorized access, damage, or theft.', 
-        'Internet Protocols': 'Rules and conventions for data exchange across networks, forming the foundational protocols for the Internet.', 
-        'Computer File Systems': 'Methods and data structures used by operating systems to store, retrieve, and organize files on storage devices.', 
-        'Distributed Computing Architecture': 'Systems that distribute computation across multiple networked computers to achieve parallel processing and resilience.', 
-        'Web Technology': 'Tools and techniques for creating, managing, and interacting with websites and web applications.', 'Programming Language Topics': 'Concepts, syntax, and semantics of programming languages, including paradigms, structures, and implementations.'
-    }, 
-    "instagram": {
-        'Normal Users': 'Individuals who use the platform for personal, non-commercial purposes. They typically have access to basic features and services.', 
-        'Commercial Users': 'Businesses or professionals utilizing the platform for commercial activities. They may have access to advanced features that support business operations.'
-    }, 
-    "reddit": {
-        'Normal Users': 'Regular users who engage with the platform without significant influence or large followings.', 
-        'Popular Users': 'Users with a substantial online presence and influence, attracting a larger audience or following.'
-    },
-    "photo": {},
-    "arxiv": {
-        "cs.AI(Artificial Intelligence)": "Covers all areas of AI except Vision, Robotics, Machine Learning, Multiagent Systems, and Computation and Language (Natural Language Processing), which have separate subject areas. In particular, includes Expert Systems, Theorem Proving (although this may overlap with Logic in Computer Science), Knowledge Representation, Planning, and Uncertainty in AI.",
-        "cs.AR(Hardware Architecture)": "Covers systems organization and hardware architecture.",
-        "cs.CC(Computational Complexity)": "Covers models of computation, complexity classes, structural complexity, complexity tradeoffs, upper and lower bounds.",
-        "cs.CE(Computational Engineering, Finance, and Science)": "Covers applications of computer science to the mathematical modeling of complex systems in the fields of science, engineering, and finance. Papers here are interdisciplinary and applications-oriented, focusing on techniques and tools that enable challenging computational simulations to be performed, for which the use of supercomputers or distributed computing platforms is often required.",
-        "cs.CG(Computational Geometry)": "Roughly includes material in ACM Subject Classes I.3.5 and F.2.2.",
-        "cs.CL(Computation and Language)": "Covers natural language processing.",
-        "cs.CR(Cryptography and Security)": "Covers all areas of cryptography and security including authentication, public key cryptosytems, proof-carrying code, etc. ",
-        "cs.CV(Computer Vision and Pattern Recognition)": "Covers image processing, computer vision, pattern recognition, and scene understanding.",
-        "cs.CY(Computers and Society)": "Covers impact of computers on society, computer ethics, information technology and public policy, legal aspects of computing, computers and education.",
-        "cs.DB(Databases)": "Covers database management, datamining, and data processing. ",
-        "cs.DC(Distributed, Parallel, and Cluster Computing)": "Covers fault-tolerance, distributed algorithms, stabilility, parallel computation, and cluster computing.",
-        "cs.DL(Digital Libraries)": "Covers all aspects of the digital library design and document and text creation.",
-        "cs.DM(Discrete Mathematics)": "Covers combinatorics, graph theory, applications of probability. ",
-        "cs.DS(Data Structures and Algorithms)": "Covers data structures and analysis of algorithms. ",
-        "cs.ET(Emerging Technologies)": "Covers approaches to information processing (computing, communication, sensing) and bio-chemical analysis based on alternatives to silicon CMOS-based technologies, such as nanoscale electronic, photonic, spin-based, superconducting, mechanical, bio-chemical and quantum technologies (this list is not exclusive). Topics of interest include (1) building blocks for emerging technologies, their scalability and adoption in larger systems, including integration with traditional technologies, (2) modeling, design and optimization of novel devices and systems, (3) models of computation, algorithm design and programming for emerging technologies.",
-        "cs.FL(Formal Languages and Automata Theory)": "Covers automata theory, formal language theory, grammars, and combinatorics on words.",
-        "cs.GL(General Literature)": "Covers introductory material, survey material, predictions of future trends, biographies, and miscellaneous computer-science related material.",
-        "cs.GR(Graphics)": "Covers all aspects of computer graphics.",
-        "cs.GT(Computer Science and Game Theory)": "Covers all theoretical and applied aspects at the intersection of computer science and game theory, including work in mechanism design, learning in games (which may overlap with Learning), foundations of agent modeling in games (which may overlap with Multiagent systems), coordination, specification and formal methods for non-cooperative computational environments. The area also deals with applications of game theory to areas such as electronic commerce.",
-        "cs.HC(Human-Computer Interaction)": "Covers human factors, user interfaces, and collaborative computing. ",
-        "cs.IR(Information Retrieval)": "Covers indexing, dictionaries, retrieval, content and analysis.",
-        "cs.IT(Information Theory)": "Covers theoretical and experimental aspects of information theory and coding. ",
-        "cs.LG(Machine Learning)": "Papers on all aspects of machine learning research (supervised, unsupervised, reinforcement learning, bandit problems, and so on) including also robustness, explanation, fairness, and methodology. cs.LG is also an appropriate primary category for applications of machine learning methods.",
-        "cs.LO(Logic in Computer Science)": "Covers all aspects of logic in computer science, including finite model theory, logics of programs, modal logic, and program verification. Programming language semantics should have Programming Languages as the primary subject area.",
-        "cs.MA(Multiagent Systems)": "Covers multiagent systems, distributed artificial intelligence, intelligent agents, coordinated interactions. and practical applications.",
-        "cs.MM(Multimedia)": "Roughly includes material in ACM Subject Class H.5.1.",
-        "cs.MS(Mathematical Software)": "",
-        "cs.NA(Numerical Analysis)": "cs.NA is an alias for math.NA. ",
-        "cs.NE(Neural and Evolutionary Computing)": "Covers neural networks, connectionism, genetic algorithms, artificial life, adaptive behavior.",
-        "cs.NI(Networking and Internet Architecture)": "Covers all aspects of computer communication networks, including network architecture and design, network protocols, and internetwork standards (like TCP/IP). Also includes topics, such as web caching, that are directly relevant to Internet architecture and performance. ",
-        "cs.OH(Other Computer Science)": "This is the classification to use for documents that do not fit anywhere else.",
-        "cs.OS(Operating Systems)": "",
-        "cs.PF(Performance)": "",
-        "cs.PL(Programming Languages)": "Covers programming language semantics, language features, programming approaches (such as object-oriented programming, functional programming, logic programming). Also includes material on compilers oriented towards programming languages; other material on compilers may be more appropriate in Architecture (AR).",
-        "cs.RO(Robotics)": "",
-        "cs.SC(Symbolic Computation)": "",
-        "cs.SD(Sound)": "Covers all aspects of computing with sound, and sound as an information channel. Includes models of sound, analysis and synthesis, audio user interfaces, sonification of data, computer music, and sound signal processing.",
-        "cs.SE(Software Engineering)": "Covers design tools, software metrics, testing and debugging, programming environments, etc. ",
-        "cs.SI(Social and Information Networks)": "Covers the design, analysis, and modeling of social and information networks, including their applications for on-line information access, communication, and interaction, and their roles as datasets in the exploration of questions in these and other domains, including connections to the social and biological sciences.",
-        "cs.SY(Systems and Control)": "cs.SY is an alias for eess.SY. This section includes theoretical and experimental research covering all facets of automatic control systems. The section is focused on methods of control system analysis and design using tools of modeling, simulation and optimization. Specific areas of research include nonlinear, distributed, adaptive, stochastic and robust control in addition to hybrid and discrete event systems. Application areas include automotive and aerospace control systems, network control, biological systems, multiagent and cooperative control, robotics, reinforcement learning, sensor networks, control of cyber-physical and energy-related systems, and control of computing systems."
-    },
-    "history": {}, 
-    "computer": {}
+    "reddit": "Reddit is also a social network where each node denotes a user, the node features are the content of users historically published subreddits, and edges denote whether two users have replied to each other. The prediction task is to classify whether a user is in the top 50% popular (average score of all subreddits).",
+    "instagram": "Instagram is a social network where edges represent following relationships, nodes represent users, and the prediction task is to classify commercial and normal users in this network.",
+    "computer": "This dataset represents a subgraph of an e-commerce network focused on computer-related items. Each node corresponds to an item in the Computer category, with node features consisting of reviews for each item. Edges denote co-purchase or co-review relationships between items. The prediction task involves classifying each item into one of 10 sub-categories, including Computer Accessories & Peripherals, Tablet Accessories, Laptop Accessories, Computers & Tablets, Computer Components, Data Storage, Networking Products, Monitors, Servers, and Tablet Replacement Parts.",
+    "photo": "This dataset represents a subgraph of an e-commerce network focused on photo-related items. Each node denotes an item in the Photo category, with node features consisting of reviews for each item. Edges denote co-purchase or co-review relationships between items. The prediction task involves classifying each item into one of 12 sub-categories, including: Video Surveillance, Accessories, Binoculars & Scopes, Video, Lighting & Studio, Bags & Cases, Tripods & Monopods, Flashes, Digital Cameras, Film Photography, Lenses, and Underwater Photography",
+    "history": "This dataset represents a subgraph of an e-commerce network focused on historical books. Each node represents a book in the History category, with node features consisting of book titles and descriptions. Edges denote co-purchase or co-review relationships between books. The prediction task involves classifying each book into one of 12 sub-categories, including: World, Americas, Asia, Military, Europe, Russia, Africa, Ancient Civilizations, Middle East, Historical Study & Educational Resources, Australia & Oceania, and Arctic & Antarctica."
 }
