+++ common/model_path.py
@@ -1,16 +1,15 @@
-common_path = "YOUR PATH OF MODELs"
-
 MODEL_PATHs = {
     # LM
-
-    "MiniLM": f"{common_path}/models--sentence-transformers--all-MiniLM-L6-v2",
-    "SentenceBert": f"{common_path}/models--sentence-transformers--multi-qa-distilbert-cos-v1/snapshots/SentenceBert",
-    "e5-large": f"{common_path}/models--intfloat--e5-large-v2/snapshots/e5-large",
-    "roberta": f"{common_path}/models--sentence-transformers--all-roberta-large-v1/snapshots/roberta",
+    "MiniLM": "sentence-transformers/all-MiniLM-L6-v2",
+    "SentenceBert": "sentence-transformers/multi-qa-distilbert-cos-v1",
+    "e5-large": "intfloat/e5-large-v2",
+    "roberta": "sentence-transformers/all-roberta-large-v1",
 
     # LLM
-    "Qwen-3B": f"{common_path}/models--Qwen--Qwen2.5-3B-Instruct/snapshots/Qwen2.5-3B", 
-    "Qwen-7B": f"{common_path}/models--Qwen--Qwen2.5-7B-Instruct/snapshots/Qwen2.5-7B",
-    "Mistral-7B": f"{common_path}/models--mistralai--Mistral-7B-Instruct-v0.2/snapshots/Mistral-7B",
-    "Llama-8B": f"{common_path}/models--meta-llama--Llama-3.1-8B-Instruct/snapshots/Llama-8B"
+    "Qwen-3B": "Qwen/Qwen2.5-3B-Instruct", 
+    "Qwen-7B": "Qwen/Qwen2.5-7B-Instruct",
+    "Qwen-14B": "Qwen/Qwen2.5-14B-Instruct",
+    "Qwen-32B": "Qwen/Qwen2.5-32B-Instruct",
+    "Mistral-7B": "mistralai/Mistral-7B-Instruct-v0.2",
+    "Llama-8B": "meta-llama/Llama-3.1-8B-Instruct",
 }
